requirements.txt
requests
psutil
pystray
pillow
pynvml
pyadl
pyinstaller
setuptools
wheel
py-cpuinfo
python-socketio
socketio
socketio-client
websocket-client
pypresence
docker
notify-py
pywebview